requirements.txt
python-telegram-bot==13.12
environs==9.5.0
redis==3.2.1
more-itertools==8.14.0
requests==2.27.1
